중복입장 버그

+++ src/main/java/psnl/liar/repository/ParticipantsRepository.java
@@ -5,4 +5,4 @@ import org.springframework.stereotype.Repository;
 import psnl.liar.entity.Participants;
 
 @Repository
-public interface ParticipantsRepository extends JpaRepository<Participants, Long> { }
+public interface ParticipantsRepository extends JpaRepository<Participants, Integer> { }
